requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-heroku==0.3.1
django-quill==0.1.0
django-quill-editor==0.1.22
django-taggit==1.5.1
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
gunicorn==20.1.0
